Use `.get` instead of indexing on config hashmaps.
The hashmap behavior was *not* doing what I expected in the case that
the key was unset or empty; instead, something internal to yaml_rust
was `expect`-ing an `Option` to be set which wasn't, and panicking. I
will need to file an issue on yaml_rust, but in the meantime... using
its `.get()` method is working just fine.