Make a New Rustacean-inspired refactor.
In the process of writing New Rustacean [e020], I realized that
the structure I was building wasn't quite right: it was treating
`Metadata` as fundamental to the `item` module, and `item`
as a child of the `builder` element, when in fact `item`s are
their own distinct concepts, and `Metadata` deserved to have
its own module. So here we are! It was also a good opportunity
to move `parse_metadata` to `Metadata::parse`, which is a far
more Rustic way of structuring this.

[e020]: http://newrustacean.com/show_notes/e020/

P.S. If this commit message seems a bit more articulate, you
can thank [The Bike Shed 105] for reminding me that this kind
of context can be incredibly helpful. Probably not so much in
*this* specific commit, but generally speaking.

[The Bike Shed 105]: http://bikeshed.fm/105