Stub out JSON Feed data type in its own crate
This is the first step in being able to generate a [JSON Feed][1]. I'd
like to come up with an actually-good API for it, and then simply do the
work to bridge to it from the main crate. I may eventually publish this
if I get it to a point where I'm happy with it. There are a handful of
related crates out there, but none which really support the full
lifecycle of *generating* a JSON Feed: only *reading* from it.