Reimplement with Comrak
This is an experiment to see if I find Comrak a suitable replacement
for pulldown-cmark. It does a number of things more nicely (footnotes
in particular... if I can get them working?!?) and the speed is good
*enough* for my purposes. Assuming I can indeed get footnotes to, uhh,
render (?) then I will likely keep this. The level of shenanigans
involved in doing syntax highlighting as part of parsing and rendering
is comparable across libraries. Different, but comparable.