A failing attempt at reworking footnotes in pulldown
This foundered on [raphlinus/pulldown-cmark#20][20], as well as a variety
of other issues, as I feared it would. The upside: I felt these muscles
work a bit, *and they work just fine*. That was a nice feeling. (Maybe I
will figure out how to actually go implement this upstream, where it
belongs anyway!)

[20]: https://github.com/raphlinus/pulldown-cmark/issues/20